folder organisation

+++ app/src/main/java/com/nestedworld/nestedworld/database/models/Monster.java
@@ -24,7 +24,7 @@ public class Monster extends SugarRecord {
     public float hp;
 
     @Expose
-    public int defense;
+    public float defense;
 
     @Expose
     public float attack;
